Refactoring out method generation-related code into its own generator class.
